requirements.txt
flask>=2.0
requests>=2.25
python-dotenv>=1.0